+++ tests/unit/core/test_wagtail_hooks.py
@@ -1,3 +1,5 @@
+import json
+
 from datetime import timedelta
 
 import pytest
@@ -207,6 +209,12 @@ def test_login_required_signup_wizard_exportplan_logged_in(domestic_site, user, 
 @pytest.mark.django_db
 def test_estimated_read_time_calculation(rf, domestic_homepage):
 
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
+
     request = rf.get('/')
     request.user = AnonymousUser()
 
@@ -242,6 +250,12 @@ def test_estimated_read_time_calculation(rf, domestic_homepage):
 @pytest.mark.django_db
 def test_estimated_read_time_calculation__checks_text_and_video(rf, domestic_homepage):
 
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
+
     request = rf.get('/')
     request.user = AnonymousUser()
 
@@ -270,7 +284,7 @@ def test_estimated_read_time_calculation__checks_text_and_video(rf, domestic_hom
     revision = detail_page.save_revision()
     revision.publish()
 
-    expected_duration = timedelta(seconds=155 + 123)  # reading + watching
+    expected_duration = timedelta(seconds=156 + 123)  # reading + watching
 
     detail_page.refresh_from_db()
     assert detail_page.estimated_read_duration != expected_duration
@@ -286,6 +300,12 @@ def test_estimated_read_time_calculation__checks_text_and_video(rf, domestic_hom
 
 @pytest.mark.django_db
 def test_estimated_read_time_calculation__checks_video(rf, domestic_homepage):
+
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
 
     request = rf.get('/')
     request.user = AnonymousUser()
@@ -323,6 +343,83 @@ def test_estimated_read_time_calculation__checks_video(rf, domestic_homepage):
 
     detail_page.refresh_from_db()
     assert detail_page.estimated_read_duration == expected_duration
+
+
+@pytest.mark.django_db
+def test_estimated_read_time_calculation__updates_only_draft_if_appropriate(
+    rf, domestic_homepage
+):
+    # This test ensures that all paths are walked for
+    # wagtail_hooks._update_data_for_appropriate_version()
+
+    # IF THIS TEST FAILS BASED ON OFF-BY-ONE-SECOND DURATIONS... check whether
+    # your changeset has slightly increased the size of the HTML page, which
+    # may have slightly pushed up the default/empty-page readtime (either in
+    # real terms or just in terms of elements that affect the calculation). If
+    # so, pushing up the expected time variables in the test is OK to do.
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    video_for_hero = make_test_video(duration=123)
+    video_for_hero.save()
+
+    detail_page = factories.DetailPageFactory(
+        parent=domestic_homepage,
+        template='learn/detail_page.html',
+        body=[],
+    )
+    assert detail_page.live is True
+
+    original_live_read_duration = detail_page.estimated_read_duration
+    assert original_live_read_duration is None
+
+    # Note: for test simplicity here, we're not adding streamfield content to our
+    # revision - it is enough to just notice how the readtimes for Draft vs Live
+    # are appropriate updated at the expected times, based on the minimal default
+    # content of a DetailPage.
+
+    revision = detail_page.save_revision()
+    assert json.loads(revision.content_json)['estimated_read_duration'] == original_live_read_duration
+
+    detail_page.refresh_from_db()
+
+    wagtail_hooks.set_read_time(
+        page=detail_page,
+        request=request
+    )
+
+    detail_page.refresh_from_db()
+
+    expected_duration = timedelta(seconds=5)  # NB just the read time of a skeleton DetailPage
+
+    # show the live version is not updated yet
+    assert detail_page.has_unpublished_changes is True
+    assert detail_page.estimated_read_duration != expected_duration
+    assert detail_page.estimated_read_duration == original_live_read_duration
+
+    # but the draft is
+    latest_rev = detail_page.get_latest_revision()
+    assert revision == latest_rev
+    assert json.loads(latest_rev.content_json)['estimated_read_duration'] == str(expected_duration)
+
+    # Now publish the draft and show it updates the live, too
+    latest_rev.publish()
+
+    detail_page.refresh_from_db()
+
+    wagtail_hooks.set_read_time(
+        page=detail_page,
+        request=request
+    )
+
+    detail_page.refresh_from_db()
+    assert detail_page.estimated_read_duration != original_live_read_duration
+
+    # NOTE: for a reason unrelated to the point of _this_ test, the readtime
+    # of the published page CAN BE calculated as slightly longer than the draft.
+    # This may be in part due to the page having a very small amount of content.
+    assert detail_page.estimated_read_duration == timedelta(seconds=5)
 
 
 @pytest.mark.django_db